add validation email user when reset password

+++ app/Services/ForgotPasswordService.php
@@ -5,6 +5,7 @@ namespace App\Services;
 use App\Mail\ResetPasswordMail;
 use App\Mail\ResetPasswordRequest;
 use App\Repositories\PasswordResetRepository;
+use App\Repositories\UserRepository;
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Support\Facades\DB;
@@ -56,15 +57,16 @@ class ForgotPasswordService
     $passwordRepository = new PasswordResetRepository();
     try{
       DB::beginTransaction();
+      $user = (new UserRepository())->getDataUserByEmail($requestedData["email"]);
+      if(!$user){
+        return false;
+      }
       $passwordRepository->deleteDataPasswordResetByEmail($requestedData["email"]);
       $reset = $passwordRepository->addNewDataPasswordReset($resetData);
-
       Mail::to($reset->email)->send(new ResetPasswordMail($reset->token));
       DB::commit();
     }catch(Exception $e){
       DB::rollback();
-
-      dd($e);
       return false;
     }
